DashTabs call getState on setState on the ContentModel
