+++ client-app/src/desktop/tabs/containers/dashwidgets/ChartWidget.js
@@ -1,0 +1,24 @@
+import {hoistCmp, creates} from '@xh/hoist/core';
+import {box} from '@xh/hoist/cmp/layout';
+import {panel} from '@xh/hoist/desktop/cmp/panel';
+import {select} from '@xh/hoist/desktop/cmp/input';
+import {chart} from '@xh/hoist/cmp/chart';
+
+import {LineChartModel} from '../../charts/LineChartModel';
+
+export const ChartWidget = hoistCmp({
+    model: creates(LineChartModel),
+    render({model}) {
+        return panel({
+            item: chart(),
+            bbar: [
+                box('Symbol: '),
+                select({
+                    bind: 'currentSymbol',
+                    options: model.symbols,
+                    enableFilter: false
+                })
+            ]
+        });
+    }
+});
