+++ client-app/src/desktop/tabs/containers/dashwidgets/GridWidget.js
@@ -1,0 +1,9 @@
+import {hoistCmp} from '@xh/hoist/core';
+
+import {sampleGrid} from '../../../common';
+
+export const GridWidget = hoistCmp({
+    render() {
+        return sampleGrid({omitGridTools: true});
+    }
+});
