+++ client-app/src/desktop/tabs/containers/dashwidgets/TreeGridWidget.js
@@ -1,0 +1,9 @@
+import {hoistCmp} from '@xh/hoist/core';
+
+import {sampleTreeGrid} from '../../../common';
+
+export const TreeGridWidget = hoistCmp({
+    render() {
+        return sampleTreeGrid({model: {includeCheckboxes: false}});
+    }
+});
