+++ client-app/src/desktop/tabs/containers/dashwidgets/index.js
@@ -1,0 +1,7 @@
+export * from './ButtonWidget';
+export * from './ButtonWidgetModel';
+export * from './ChartWidget';
+export * from './GridWidget';
+export * from './GridWidgetModel';
+export * from './PanelWidget';
+export * from './TreeGridWidget';
